move non-table mutations into their own directory, move more generic functions into helpers, implement add_pipeline_1 step
